moved Logger.java to utils/Logger.java
