+++ STYLE_GUIDE.md
@@ -40,3 +40,4 @@ public class ConventionsHelloWorld {
 	}
 }
 ```
+
