+++ src/cubyz/rendering/MainRenderer.java
@@ -12,7 +12,7 @@ import org.joml.Vector3d;
 import org.joml.Vector3f;
 import org.joml.Vector4f;
 
-import cubyz.Logger;
+import cubyz.utils.Logger;
 import cubyz.client.BlockMeshes;
 import cubyz.client.ChunkMesh;
 import cubyz.client.ClientSettings;
