+++ src/cubyz/rendering/rotation/StackableRotation.java
@@ -8,7 +8,7 @@ import org.joml.Vector3f;
 import org.joml.Vector3i;
 import org.joml.Vector4d;
 
-import cubyz.Logger;
+import cubyz.utils.Logger;
 import cubyz.api.Resource;
 import cubyz.client.BlockMeshes;
 import cubyz.rendering.models.CubeModel;
