+++ src/cubyz/utils/Logger.java
@@ -1,4 +1,4 @@
-package cubyz;
+package cubyz.utils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
